Ajout des tests CA minimum et aléatoire
