--- src/main/tests/CommercialTest.java
@@ -1,22 +1,0 @@
-import org.assertj.core.api.Assertions;
-import org.junit.Test;
-
-import com.ipiecoles.java.java340.model.Commercial;
-
-public class CommercialTest {
-	
-	@Test
-	public void testerPrimeAnnuelleWithCANull() {
-		// Given
-		Commercial commercial = new Commercial(); 
-		commercial.setCaAnnuel(null);
-		
-		// When 
-		Double prime = commercial.getPrimeAnnuelle();
-		
-		// Then
-		Assertions.assertThat(prime).isEqualTo(500d);
-		
-	}
-
-}
